+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Modal window when closing ESI fleets manually
 - Ability to re-open FAT links for a certain time after they have expired
   (`manage_afat` permissions are needed to re-open FAT links)
+- Logs view (`log_view` permissions are needed to view the logs)
 
 ### Changed